Refactor: View in multi part
