+++ SDKAndroid/src/main/java/app/cybrid/sdkandroid/components/kyc/compose/KYCView_Loading.kt
@@ -1,0 +1,54 @@
+package app.cybrid.sdkandroid.components.kyc.compose
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import app.cybrid.sdkandroid.R
+import app.cybrid.sdkandroid.core.Constants
+import app.cybrid.sdkandroid.ui.Theme.robotoFont
+
+@Composable
+fun KYCView_Loading() {
+
+    Box(
+        modifier = Modifier
+            .height(120.dp)
+            .testTag(Constants.IdentityVerificationView.LoadingView.id)
+    ) {
+        Column(
+            modifier = Modifier.fillMaxSize(),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(
+                text = stringResource(id = R.string.kyc_view_loading_text),
+                fontFamily = robotoFont,
+                fontWeight = FontWeight.Normal,
+                fontSize = 17.sp,
+                color = colorResource(id = R.color.primary_color)
+            )
+            CircularProgressIndicator(
+                modifier = Modifier
+                    .padding(top = 16.dp)
+                    .testTag(Constants.IdentityVerificationView.LoadingViewIndicator.id),
+                color = colorResource(id = R.color.primary_color)
+            )
+        }
+    }
+}
+
+@Preview(showBackground = true, widthDp = 200, heightDp = 300)
+@Composable
+fun KYCView_Loading_Preview() {
+    KYCView_Loading()
+}
