+++ SDKAndroid/src/main/res/raw/test_create_customer_success.json
@@ -1,0 +1,1 @@
+{"guid":"1234","type":"individual","created_at":"2022-06-23T07:08:16.718Z","state":"storing"}
